put message/connection in their own files
